API/requirements_flask.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0

# MQTT
paho-mqtt==2.1.0

# Environment & Config
python-dotenv==1.0.0

# Validation
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0

# HTTP Requests
requests==2.31.0

# Development
python-decouple==3.8

# Additional utilities
email-validator==2.1.0
numpy==2.2.3